Streamline code architecture
