+++ MathsEngine/MathsEngine.csproj
@@ -49,15 +49,14 @@
     <Compile Include="Core\Menu\Pure\MatrixMenu.cs" />
     <Compile Include="Core\Menu\Pure\TrigonometryMenu.cs" />
     <Compile Include="Core\Menu\Statistics\DispersionMenu.cs" />
-    <Compile Include="Modules\Core\PureHelpers\PythagorasTheorem.cs" />
-    <Compile Include="Modules\Core\PureHelpers\Trigonometry.cs" />
-    <Compile Include="Modules\Core\StatisticsHelpers\AverageCalculator.cs" />
+    <Compile Include="Modules\Pure\PythagorasTheorem\PythagorasTheorem.cs" />
+    <Compile Include="Modules\Pure\Trigonometry\Trigonometry.cs" />
+    <Compile Include="Modules\Statistics\AverageCalculator.cs" />
     <Compile Include="Core\Menu\Pure\PythagorasMenu.cs" />
     <Compile Include="Modules\Mechanics\Dynamics\NewtonsLawsCalculator.cs" />
     <Compile Include="Modules\Mechanics\UniformAcceleration\UniformAccelerationCalculator.cs" />
     <Compile Include="Modules\Pure\Matrices\MatrixBase.cs" />
     <Compile Include="Modules\Pure\Matrices\MatrixCalculator.cs" />
-    <Compile Include="Modules\Pure\PythagorasTheorem\PythagorasTheorem.cs" />
     <Compile Include="Modules\Pure\Trigonometry\SideType.cs" />
     <Compile Include="Modules\Statistics\BivariateAnalysis\BivariateAnalysisCalculator.cs" />
     <Compile Include="Modules\Statistics\BivariateAnalysis\BivariateAnalysis.cs" />
